+++ src/Common/AbstractEntity.php
@@ -152,12 +152,17 @@ abstract class AbstractEntity implements \JsonSerializable, EntityInterface
      * Read method (GET).
      *
      * @param array $params
+     * @param array $queryStrings
+     * @param bool  $shouldTheExpectedResponseBeMappedOntoTheEntity
      *
      * @return mixed
      * @throws \Exception
      */
-    public function read(array $params = [])
-    {
+    public function read(
+        array $params = [],
+        array $queryStrings = [],
+        bool $shouldTheExpectedResponseBeMappedOntoTheEntity = true
+    ) {
         $method = 'get';
         $class  = get_called_class();
         $entity = new $class($this->manager);
@@ -165,10 +170,14 @@ abstract class AbstractEntity implements \JsonSerializable, EntityInterface
         $customHeaders = $this->getHeaders()['read'];
         $header        = $this->manager->getHeader($customHeaders);
 
-        $uri      = $this->manager->getEntityUri($entity, $method, $params);
+        $uri      = $this->manager->getEntityUri($entity, $method, $params, $queryStrings);
         $response = $this->manager->execute($entity, $uri, $method, $header);
         $this->obfuscateAuthorizationHeader($header);
-        return $this->manager->handleResponse($response, $method, $entity);
+        return $this->manager->handleResponse(
+            $response,
+            $method,
+            $shouldTheExpectedResponseBeMappedOntoTheEntity ? $entity : null
+        );
     }
 
     /**
@@ -194,17 +203,18 @@ abstract class AbstractEntity implements \JsonSerializable, EntityInterface
      * Save method with params (POST).
      *
      * @param array $params
+     * @param array $queryStrings
      *
      * @return mixed
      * @throws \Exception
      */
-    public function saveWithParams(array $params = [])
+    public function saveWithParams(array $params = [], array $queryStrings = [])
     {
         $method = 'post';
         $customHeaders = $this->getHeaders()['save'];
         $header        = $this->manager->getHeader($customHeaders);
 
-        $uri      = $this->manager->getEntityUri($this, $method, $params);
+        $uri      = $this->manager->getEntityUri($this, $method, $params, $queryStrings);
         $response = $this->manager->execute($this, $uri, $method, $header);
         $this->obfuscateAuthorizationHeader($header);
         return $this->manager->handleResponse($response, $method);
@@ -235,7 +245,7 @@ abstract class AbstractEntity implements \JsonSerializable, EntityInterface
      */
     public function obfuscateAuthorizationHeader(array $headers)
     {
-        Sdk::$cache['last_headers'] = preg_replace('/(Authorization: Bearer) (.*)/i', '$1 xxx', $headers);
+        Sdk::$cache['last_headers'] = preg_replace('/(Authorization:) (.*)/i', '$1 xxx', $headers);
     }
 
     /**