+++ src/Sdk.php
@@ -15,6 +15,7 @@ use MercadoPago\PP\Sdk\Entity\Payment\Payment;
 use MercadoPago\PP\Sdk\Entity\Payment\PaymentV2;
 use MercadoPago\PP\Sdk\Entity\Payment\PaymentV21;
 use MercadoPago\PP\Sdk\Entity\Preference\Preference;
+use MercadoPago\PP\Sdk\Entity\PaymentMethods\PaymentMethods;
 use MercadoPago\PP\Sdk\HttpClient\HttpClient;
 use MercadoPago\PP\Sdk\HttpClient\Requester\CurlRequester;
 use MercadoPago\PP\Sdk\HttpClient\Requester\RequesterInterface;
@@ -43,12 +44,14 @@ class Sdk
      * @param String $platform_id
      * @param String $product_id
      * @param String $integrator_id
+     * @param String $public_key
      */
     public function __construct(
         string $access_token,
         string $platform_id,
         string $product_id,
-        string $integrator_id
+        string $integrator_id,
+        string $public_key
     ) {
         $this->requester = new CurlRequester();
         $this->config = new Config();
@@ -56,6 +59,7 @@ class Sdk
         $this->config->__set('platform_id', $platform_id);
         $this->config->__set('product_id', $product_id);
         $this->config->__set('integrator_id', $integrator_id);
+        $this->config->__set('public_key', $public_key);
     }
 
     /**
@@ -149,5 +153,16 @@ class Sdk
     public function getRegisterErrorLogInstance()
     {
         return $this->getEntityInstance('MercadoPago\PP\Sdk\Entity\Monitoring\RegisterErrorLog', Constants::BASEURL_MP);
+    }
+
+    /**
+     * @return PaymentMethods
+    */
+    public function getPaymentMethodsInstance()
+    {
+        return $this->getEntityInstance(
+            'MercadoPago\PP\Sdk\Entity\PaymentMethods\PaymentMethods',
+            Constants::BASEURL_MP
+        );
     }
 }