+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## V1.13.0
+### Added
+- `PaymentMethods` entity to get payment methods using the Core Payment Methods service API
+- Create `getPaymentMethods` and `getPaymentMethodsByGroupBy` methods in `PaymentMethods` class
+- Integration tests for PaymentMethods scenario
+- Adds the possibility of passing query strings in requests
+- Create `setHeadersAsKeyAndValueMap`, `isHeadersAsKeyAndValueMap`, and `normalizeHeaders` methods in Manager class
 
 ## V1.12.0
 ### Added